Add 验证注解事务是否有效

+++ youth/src/main/java/com/xurui/youth/mapper/UserMapper.java
@@ -7,7 +7,7 @@ import org.apache.ibatis.annotations.Mapper;
 public interface UserMapper {
     int deleteByPrimaryKey(Long id);
 
-    int insert(User record);
+    long insert(User record);
 
     int insertSelective(User record);
